+++ dev/src/js/dev/src/ts/states/game.js
@@ -20,21 +20,33 @@ var GameState = (function (_super) {
     }
     GameState.prototype.create = function () {
         this.game.physics.lite.enablePhysics();
-        var texture = Lightning.Geometry.Rect(20, 20).generateTexture();
+        var texture = Lightning.Geometry.Rect(5, 5).generateCanvasTexture();
         var pool = this.game.physics.lite.createPool('test');
-        for (var i = 0; i < 50; i++) {
+        for (var i = 0; i < 1; i++) {
             var sprite = new Lightning.Sprite();
             sprite.texture = texture;
-            sprite.x = Lightning.Maths.rngInt(0, this.game.width);
-            sprite.y = Lightning.Maths.rngInt(0, this.game.height);
-            sprite.setAnchor(0.5);
+            sprite.x = Lightning.Maths.rngFloat(0, this.game.width);
+            sprite.y = Lightning.Maths.rngFloat(0, this.game.height);
+            // sprite.rotation = Lightning.Maths.rngFloat(0, 12);
             this.add(sprite);
             sprite.enablePhysicsBody();
             pool.add(sprite.body);
             sprite.body.enableDebug();
-            sprite.body.velocity.x = Lightning.Maths.rngInt(-2, 2);
-            sprite.body.velocity.y = Lightning.Maths.rngInt(-2, 2);
+            sprite.body.velocity.x = Lightning.Maths.rngFloat(-5, 5);
+            sprite.body.velocity.y = Lightning.Maths.rngFloat(-5, 5);
+            sprite.body.collideOnWorldBounds = true;
+            sprite.body.gravityEnabled = true;
+            sprite.body.restitution = 0.5;
         }
+        var bigGuy = new Lightning.Sprite();
+        bigGuy.texture = Lightning.Geometry.Rect(50, 50).generateCanvasTexture();
+        bigGuy.x = this.game.width / 2;
+        bigGuy.y = this.game.height / 2;
+        this.add(bigGuy);
+        bigGuy.enablePhysicsBody();
+        bigGuy.body.enableDebug();
+        var cEvent = this.game.physics.lite.createCollisionEvent('t', bigGuy.body, pool.bodies);
+        cEvent.onCollide(this.onCollide, this);
         /**
         1.  * creating basic tween
          */
@@ -190,6 +202,8 @@ var GameState = (function (_super) {
         //     customStorage.setItem('test', 67890);
         //     customStorage.getItem('test');
         // ... //
+    };
+    GameState.prototype.onCollide = function (obj1, obj2) {
     };
     return GameState;
 }(Lightning.State));