+++ dev/src/js/states/menu.js
@@ -16,8 +16,17 @@ var MenuState = (function (_super) {
         return _super !== null && _super.apply(this, arguments) || this;
     }
     MenuState.prototype.create = function () {
-        this.game.states.destroy('preload');
-        this.game.states.start('game');
+        var _this = this;
+        console.log(this.game.world.children);
+        var button = Lightning.Geometry.Rect(50, 50);
+        button.tint = 0xff22aa;
+        this.add(button);
+        button.interactive = true;
+        button.on('mousedown', function () {
+            _this.game.states.start('game');
+        });
+    };
+    MenuState.prototype.update = function () {
     };
     return MenuState;
 }(Lightning.State));