+++ dev/src/ts/states/game.ts
@@ -8,6 +8,11 @@ export default class GameState extends Lightning.State {
      private _start:boolean = false;
 
     create() {
+        console.log(this.game.world.children);
+
+        setTimeout(() => {
+            this.game.states.start('menu');
+        }, 1500);
 
         this.game.physics.lite.enablePhysics();
         let texture:Lightning.Texture =  Lightning.Geometry.Rect(5, 5).generateCanvasTexture();
@@ -27,6 +32,9 @@ export default class GameState extends Lightning.State {
 
             sprite.body.velocity.x = Lightning.Maths.rngFloat(-5, 5);
             sprite.body.velocity.y = Lightning.Maths.rngFloat(-5, 5);
+            sprite.body.collideOnWorldBounds = true;
+            sprite.body.gravityEnabled = true;
+            sprite.body.restitution = 0.5;
         }
 
         let bigGuy = new Lightning.Sprite();
@@ -37,17 +45,11 @@ export default class GameState extends Lightning.State {
 
         bigGuy.enablePhysicsBody();
         bigGuy.body.enableDebug();
+        bigGuy.body.static = true;
 
         let cEvent = this.game.physics.lite.createCollisionEvent('t', bigGuy.body, pool.bodies);
-        cEvent.onCollide((obj1:Lightning.LitePhysicsBody, obj2:Lightning.LitePhysicsBody) => {
-            obj2.velocity.x *= -1;
-            obj2.velocity.y *= -1;
-            obj2.pauseCollisionDetection = true;
 
-            setTimeout(() => {
-                obj2.pauseCollisionDetection = false;
-            }, 20);
-        });
+        cEvent.onCollide(this.onCollide, this);
 
         
         /**
@@ -249,5 +251,9 @@ export default class GameState extends Lightning.State {
     //     customStorage.getItem('test');
         // ... //
       
+    }
+
+    private onCollide(obj1, obj2) {
+        console.log('hi')
     }
 }