+++ dev/src/ts/states/menu.ts
@@ -1,7 +1,18 @@
 export default class MenuState extends Lightning.State {
 
     create() {
-        this.game.states.destroy('preload');
-        this.game.states.start('game');
+        console.log(this.game.world.children);
+        let button:Lightning.Graphics = Lightning.Geometry.Rect(50, 50);
+        button.tint = 0xff22aa;
+        this.add(button);
+
+        button.interactive = true;
+        button.on('mousedown', () => {
+            this.game.states.start('game');
+        });
+    }
+
+    update() {
+
     }
 }